# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AppleLoginRequest:
    properties:
      fullName:
        properties:
          familyName:
            type: string
          givenName:
            type: string
        type: object
      identityToken:
        type: string
    required:
    - identityToken
    type: object
  models.CreateReviewRequest:
    properties:
      comment:
        type: string
      foods:
        items:
          $ref: '#/definitions/models.ReviewFoodItem'
        type: array
      imageURL:
        type: string
      mealTime:
        type: string
      name:
        type: string
      rating:
        type: integer
      speed:
        type: string
    required:
    - foods
    - mealTime
    - name
    - rating
    - speed
    type: object
  models.CreateStandardFoodRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      imageURL:
        type: string
      name:
        type: string
      parents:
        items:
          type: string
        type: array
      speed:
        type: string
    required:
    - categories
    - imageURL
    - name
    - parents
    - speed
    type: object
  models.GoogleLoginRequest:
    properties:
      idToken:
        type: string
    required:
    - idToken
    type: object
  models.KakaoLoginRequest:
    properties:
      accessToken:
        type: string
    required:
    - accessToken
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      accessToken:
        type: string
      isNewUser:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.MainFeedResponse:
    properties:
      food:
        $ref: '#/definitions/models.StandardFood'
      isLiked:
        type: boolean
    type: object
  models.ResolveFoodItemsRequest:
    properties:
      names:
        items:
          type: string
        type: array
    required:
    - names
    type: object
  models.Review:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      day:
        type: integer
      foods:
        items:
          $ref: '#/definitions/models.ReviewFoodItem'
        type: array
      id:
        type: string
      imageURL:
        type: string
      mealTime:
        type: string
      name:
        type: string
      rating:
        type: integer
      speed:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  models.ReviewFoodItem:
    properties:
      foodID:
        type: string
      foodName:
        type: string
      type:
        type: string
    type: object
  models.SignUpRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.StandardFood:
    properties:
      categories:
        items:
          type: string
        type: array
      id:
        type: string
      imageURL:
        type: string
      likeCount:
        type: integer
      name:
        type: string
      parents:
        items:
          type: string
        type: array
      reviewCount:
        type: integer
      speed:
        type: string
      totalRating:
        type: integer
    required:
    - imageURL
    - name
    - speed
    type: object
  models.UpdateReviewRequest:
    properties:
      comment:
        type: string
      imageURL:
        type: string
      mealTime:
        type: string
      rating:
        type: integer
    required:
    - mealTime
    - rating
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      day:
        type: integer
      email:
        type: string
      id:
        type: string
      loginMethod:
        type: string
      username:
        type: string
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorDetail'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: 밥땡의 백엔드 서버 API 명세서
  title: Bobttaeng API Server
  version: "1.0"
paths:
  /admin/standard-foods:
    post:
      consumes:
      - application/json
      description: 관리자 권한으로 새로운 표준 음식들을 생성한다.
      parameters:
      - description: 음식 정보 목록
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/models.CreateStandardFoodRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: 음식 생성 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StandardFood'
                  type: array
              type: object
      summary: 표준 음식 생성
      tags:
      - Admin
  /auth/apple:
    post:
      consumes:
      - application/json
      description: 애플 아이덴티티 토큰으로 로그인한다.
      parameters:
      - description: 애플 아이덴티티 토큰 및 전체 이름
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AppleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공 정보
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
      summary: 애플 로그인
      tags:
      - Auth
  /auth/check-username:
    get:
      consumes:
      - application/json
      description: 입력한 아이디(Username)가 이미 존재하는지 확인한다.
      parameters:
      - description: 확인할 아이디
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '중복 여부 (true: 존재함)'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: boolean
              type: object
      summary: 아이디 중복 확인
      tags:
      - Auth
  /auth/google:
    post:
      consumes:
      - application/json
      description: 구글 ID 토큰으로 로그인한다.
      parameters:
      - description: 구글 ID 토큰
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공 정보
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
      summary: 구글 로그인
      tags:
      - Auth
  /auth/kakao:
    post:
      consumes:
      - application/json
      description: 카카오 액세스 토큰으로 로그인한다.
      parameters:
      - description: 카카오 액세스 토큰
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.KakaoLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공 정보
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
      summary: 카카오 로그인
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 아이디와 비밀번호로 로그인한다.
      parameters:
      - description: 로그인 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공 정보
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "401":
          description: 아이디 또는 비밀번호 불일치
          schema:
            $ref: '#/definitions/response.Response'
      summary: 일반 로그인
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: 아이디와 비밀번호로 새로운 유저를 등록한다.
      parameters:
      - description: 회원가입 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 성공 메시지
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "409":
          description: 이미 존재하는 아이디
          schema:
            $ref: '#/definitions/response.Response'
      summary: 일반 회원가입
      tags:
      - Auth
  /foods/{foodID}:
    get:
      consumes:
      - application/json
      description: 음식 ID를 통해 표준 음식 정보를 조회한다.
      parameters:
      - description: 음식 ID
        in: path
        name: foodID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 음식 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StandardFood'
              type: object
        "400":
          description: 잘못된 ID 형식
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 음식을 찾을 수 없음
          schema:
            $ref: '#/definitions/response.Response'
      summary: 표준 음식 조회
      tags:
      - Food
  /foods/{foodID}/likes:
    delete:
      consumes:
      - application/json
      description: 특정 음식의 좋아요를 취소한다.
      parameters:
      - description: 음식 ID
        in: path
        name: foodID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 성공 메시지
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: 좋아요 기록 없음
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 음식 좋아요 취소
      tags:
      - Like
    post:
      consumes:
      - application/json
      description: 특정 음식에 좋아요를 표시한다.
      parameters:
      - description: 음식 ID
        in: path
        name: foodID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 성공 메시지
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 잘못된 음식 ID 형식
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: 이미 좋아요한 음식
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 음식 좋아요
      tags:
      - Like
  /foods/main-feed:
    get:
      consumes:
      - application/json
      description: 유저에게 추천한 기록을 바탕으로 카뉴 음식 목록을 가져온다.
      parameters:
      - description: 속도 (fast/slow)
        in: query
        name: speed
        required: true
        type: string
      - description: 조회 개수 (최대 10개)
        in: query
        name: count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MainFeedResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 카뉴 음식 조회
      tags:
      - Food
  /foods/resolve:
    post:
      consumes:
      - application/json
      description: 음식 이름들을 받아 리뷰용 음식 항목들을 생성 및 반환한다.
      parameters:
      - description: 음식 이름 목록
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ResolveFoodItemsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 음식 항목 반환 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ReviewFoodItem'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 리뷰용 음식 반환
      tags:
      - Food
  /reviews:
    post:
      consumes:
      - application/json
      description: 새로운 리뷰를 생성한다.
      parameters:
      - description: 리뷰 정보
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 리뷰 생성 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Review'
              type: object
        "400":
          description: 잘못된 요청 본문
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 리뷰 생성
      tags:
      - Review
  /reviews/{reviewID}:
    patch:
      consumes:
      - application/json
      description: 기존 리뷰 정보를 수정한다.
      parameters:
      - description: 리뷰 ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: 리뷰 수정 정보
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 리뷰 수정 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Review'
              type: object
        "403":
          description: 수정 권한 없음
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 리뷰 수정
      tags:
      - Review
  /users/me:
    get:
      consumes:
      - application/json
      description: 현재 로그인한 유저 정보를 가져온다.
      produces:
      - application/json
      responses:
        "200":
          description: 현재 유저 정보
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      security:
      - BearerAuth: []
      summary: 내 정보 조회
      tags:
      - User
  /users/me/liked-foods:
    get:
      consumes:
      - application/json
      description: 현재 유저가 좋아요한 음식 목록을 조회한다.
      produces:
      - application/json
      responses:
        "200":
          description: 좋아요 음식 목록
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StandardFood'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 좋아요한 음식 목록 조회
      tags:
      - User
  /users/me/reviews:
    get:
      consumes:
      - application/json
      description: 특정 날짜에 작성한 리뷰 목록을 조회한다.
      parameters:
      - description: 조회할 날짜
        in: query
        name: day
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 리뷰 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Review'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 날짜별 리뷰 조회
      tags:
      - Review
  /users/me/sync-day:
    post:
      consumes:
      - application/json
      description: 가입일로부터의 경과 일수를 바탕으로 유저의 Day를 업데이트한다.
      produces:
      - application/json
      responses:
        "200":
          description: 동기화 성공 메시지
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Day 동기화
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: '"Bearer {token}" 형식으로 JWT 토큰을 전달'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"